requirements.txt
bs4
selenium
tweepy
utils-nuuuwan
gig-nuuuwan
geo-nuuuwan
